requirements.txt
greenlet==3.0.3
numpy==1.24.4
pandas==2.0.3
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.12.1
tzdata==2024.1
